summary-barchart done.
