+++ src/components/powerInfoPage/index.tsx
@@ -4,7 +4,7 @@ import Chart from './chart';
 import ModeButton from './modeButton';
 import DatePicker from './datePicker';
 // import Summary from './summary';
-import Summary from './summary/';
+import Summary from './summary';
 import Table from './table';
 
 const PowerInfoPageContainer: React.FC = () => {
