--- src/scss/components/powerInfoPage/_summary.scss
@@ -1,36 +1,0 @@
-.powerinfo-summary-container {
-  width: 100%;
-  height: 100%;
-  overflow-y: hidden;
-  overflow-x: hidden;
-}
-
-.powerinfo-summary-svg {
-  width: 100%;
-  height: 100%;
-  border-radius: 5px;
-}
-
-// 1920 < size
-@media all and (min-width: 1920px) {
-// ignore
-}
-
-// 1280 < size <= 1920
-@media all and (min-width: 1280px) and (max-width: 1920px) {
-
-}
-
-// 720 < size <= 1280
-@media all and (min-width: 720px) and (max-width: 1280px) {
-
-}
-
-// 320 < size <= 720
-@media all and (min-width: 320px) and (max-width: 720px) {
-
-}
-
-@media all and (max-width: 320px) {
-
-}
