+++ src/scss/components/powerInfoPage/_table.scss
@@ -33,26 +33,31 @@ hr {
 
 .powerinfo-table-title-date {
   width: 20%;
+  // border: red solid 1px;
 }
 
 .powerinfo-table-title-time {
   text-align: center;
   width: 17%;
+  // border: red solid 1px;
 }
 
 .powerinfo-table-title-value {
   text-align: center;
   width: 17%;
+  // border: red solid 1px;
 }
 
 .powerinfo-table-title-type {
   text-align: center;
   width: 17%;
+  // border: red solid 1px;
 }
 
 .powerinfo-table-title-url {
   text-align: center;
   width: 17%;
+  // border: red solid 1px;
 }
 
 .powerinfo-table-list-item-container {
