+++ src/scss/components/powerInfoPage/summary/_barChart.scss
@@ -1,0 +1,78 @@
+.powerinfo-summary-barchart-container {
+  margin: 10%;
+  width: 80%;
+  height: 80%;
+  display: flex;
+  flex-wrap: wrap;
+  color: #707070;
+  // border: blue solid 1px;
+}
+
+.powerinfo-summary-barchart-date {
+  flex-basis: 100%;
+  height: 10%;
+  text-align: center;
+  font-weight: bold;
+  // border: gray solid 1px;
+}
+
+.powerinfo-summary-barchart-mode {
+  flex-basis: 100%;
+  margin-left: 0%;
+  height: auto;
+  text-align: center;
+  font-weight: bold;
+  overflow: auto;
+  // border: gray solid 1px;
+  >span {
+    font-weight: lighter;
+  }
+}
+
+.powerinfo-summary-barchart-svg {
+  flex-basis: 100%;
+  height: 80%;
+  border-radius: 5px;
+  // border: gray solid 1px;
+}
+
+// 1920 < size
+@media all and (min-width: 1920px) {
+  .powerinfo-summary-barchart-container {
+    font-size: 22px;
+    margin: 10%;
+    width: 80%;
+    height: 80%;
+  }
+}
+
+// 1280 < size <= 1920
+@media all and (min-width: 1280px) and (max-width: 1920px) {
+  .powerinfo-summary-barchart-container {
+    font-size: 16px;
+  }
+}
+
+// 720 < size <= 1280
+@media all and (min-width: 720px) and (max-width: 1280px) {
+  .powerinfo-summary-barchart-container {
+    font-size: 14px;
+  }
+}
+
+// 320 < size <= 720
+@media all and (min-width: 320px) and (max-width: 720px) {
+  .powerinfo-summary-barchart-container {
+    font-size: 12px;
+  }
+}
+
+// size <= 320
+@media all and (max-width: 320px) {
+  .powerinfo-summary-barchart-container {
+    font-size: 10px;
+    margin: 1%;
+    width: 98%;
+    height: 98%;
+  }
+}
