+++ src/scss/components/powerInfoPage/summary/_index.scss
@@ -5,6 +5,6 @@
   width: 100%;
   height: 100%;
   background-color: white;
-  border: pink solid 3px;
-  overflow: auto;
+  border: white solid 3px;
+  // overflow: auto;
 }
