+++ src/scss/components/powerInfoPage/summary/_list.scss
@@ -7,14 +7,15 @@
   height: 80%;
   background-color: white;
   font-size: 16px;
-  border: green solid 2px;
-  overflow-y: auto;
+  // border: green solid 2px;
+  overflow-y: hidden;
 }
 
 .powerinfo-summary-list-date {
   flex-basis: 100%;
   height: 10%;
   text-align: center;
+  align-self: flex-end;
   color: #707070;
   font-weight: bold;
   // font-size: 14px;
@@ -71,7 +72,7 @@
 
 .powerinfo-summary-list-posdata-item-container {
   flex-basis: 70%;
-  height: 15%;
+  height: auto;
   margin-left: 15%;
   margin-right: 15%;
   align-self: center;
@@ -92,7 +93,7 @@
 
 .powerinfo-summary-list-negdata-item-container {
   flex-basis: 70%;
-  height: 15%;
+  height: auto;
   margin-left: 15%;
   margin-right: 15%;
   align-self: center;
@@ -131,6 +132,7 @@
     text-align: start;
     // border: red solid 1px;
   }
+  // border: red solid 1px;
 }
 
 // 1920 < size
@@ -144,20 +146,30 @@
 @media all and (min-width: 1280px) and (max-width: 1920px) {
   .powerinfo-summary-list-container {
     font-size: 16px;
+    margin: 5%;
+    margin-bottom: 0;
+    width: 90%;
+    height: 95%;
   }
 }
 
 // 720 < size <= 1280
 @media all and (min-width: 720px) and (max-width: 1280px) {
   .powerinfo-summary-list-container {
-    font-size: 14px;
+    font-size: 12px;
+    margin: 5%;
+    width: 90%;
+    height: 90%;
   }
 }
 
 // 320 < size <= 720
 @media all and (min-width: 320px) and (max-width: 720px) {
   .powerinfo-summary-list-container {
-    font-size: 12px;
+    font-size: 10px;
+    margin: 1%;
+    width: 98%;
+    height: 98%;
   }
 }
 
@@ -165,5 +177,46 @@
 @media all and (max-width: 320px) {
   .powerinfo-summary-list-container {
     font-size: 10px;
+    margin: 1%;
+    width: 98%;
+    height: 98%;
+  }
+
+  .powerinfo-summary-list-date {
+    flex-basis: 45%;
+    margin-left: 5%;
+    height: auto;
+    text-align: start;
+    // border: green solid 1px;
+  }
+
+  .powerinfo-summary-list-mode {
+    flex-basis: 50%;
+    height: auto;
+    text-align: start;
+    // border: green solid 1px;
+  }
+
+  .powerinfo-summary-list-data-container {
+    align-self: center;
+    height: 75%;
+  }
+
+  .powerinfo-summary-list-posdata-item-container {
+    flex-basis: 45%;
+    margin-left: 5%;
+    margin-right: 0%;
+    // border: green solid 1px;
+  }
+
+  .powerinfo-summary-list-negdata-item-container {
+    flex-basis: 45%;
+    margin-left: 5%;
+    margin-right: 0%;
+  }
+
+  .powerinfo-summary-list-demand-container {
+    height: auto;
+    // border: green solid 1px;
   }
 }
